test_suite.py cleans up the tests (if any) before starting them
